+++ packages/sdk-swift/CHANGELOG.md
@@ -4,7 +4,9 @@ All notable changes to `relaycast-swift` will be documented in this file.
 
 See the [root changelog](../../CHANGELOG.md) for cross-package release highlights.
 
-## [Unreleased]
+## [Unreleased - Patch]
+
+- Allowed agent models to decode the hosted lifecycle statuses used during realtime connection setup.
 
 ## [6.0.0] - 2026-07-09